refactor(components): organise helper functions in separate files
