+++ src/components/shared/helpers/createBlurHandler.js
@@ -1,0 +1,9 @@
+const createBlurHandler = (input, onBlur) => (payload, event) => {
+    input.onBlur(event)
+
+    if (onBlur && typeof onBlur === 'function') {
+        onBlur(payload, event)
+    }
+}
+
+export { createBlurHandler }
