+++ src/components/shared/helpers/createFocusHandler.js
@@ -1,0 +1,9 @@
+const createFocusHandler = (input, onFocus) => (payload, event) => {
+    input.onFocus(event)
+
+    if (onFocus && typeof onFocus === 'function') {
+        onFocus(payload, event)
+    }
+}
+
+export { createFocusHandler }
