+++ src/components/shared/helpers/getValidationText.js
@@ -1,0 +1,6 @@
+import { hasError } from './hasError.js'
+
+const getValidationText = (meta, validationText, error) =>
+    validationText || (hasError(meta, error) && meta.error) || ''
+
+export { getValidationText }
