+++ src/components/shared/helpers/hasError.js
@@ -1,0 +1,3 @@
+const hasError = (meta, error) => error || (meta.touched && meta.invalid)
+
+export { hasError }
