+++ src/components/shared/helpers/isLoading.js
@@ -1,0 +1,4 @@
+const isLoading = (meta, loading, showLoadingStatus) =>
+    loading || (showLoadingStatus && meta.validating)
+
+export { isLoading }
