+++ src/components/shared/helpers/isValid.js
@@ -1,0 +1,4 @@
+const isValid = (meta, valid, showValidStatus) =>
+    valid || (showValidStatus && meta.touched && meta.valid)
+
+export { isValid }
